fix: remove repeated code for sprite offset
There was some code which calculated the render destination rectangle based on an entity's transform and sprite components.

The destination rectangle was determined by taking the sprite's position (transform) and applying an offset based on its size (sprite).

That code was repeated for at least the sprite render and bounding box render calls - and so was pulled out into its own function which can now be called repeatedly.

This is only part of the puzzle as ultimately it would be useful to be able to determine the 'bottom middle
' of an SDL_Rect - so more thought is needed on how to encapsulate that.
