+++ src/engine/systems/render.h
@@ -9,6 +9,29 @@
 #include "sprite.h"
 #include "constants.h"
 
+SDL_Rect get_render_target(const Transform& transform, const Sprite& sprite, const SDL_Rect& camera) {
+    return SDL_Rect {
+        (static_cast<int>(transform.position.x) + sprite.horizontal_offset_px) - camera.x, 
+        (static_cast<int>(transform.position.y) + sprite.vertical_offset_px) - camera.y, 
+        sprite.width_px,
+        sprite.height_px
+    };
+}
+
+void render_bounding_box(const entt::registry& registry, const SDL_Rect& camera, SDL_Renderer* renderer) {
+    auto vertical_tiles = registry.view<Transform, VerticalSprite>();
+    vertical_tiles.use<Transform>();
+    for (auto entity: vertical_tiles) {
+
+        const Transform& transform {vertical_tiles.get<Transform>(entity)};
+        const VerticalSprite& sprite {vertical_tiles.get<VerticalSprite>(entity)};
+
+        SDL_Rect bounding_box {get_render_target(transform, sprite, camera)};
+                
+        SDL_SetRenderDrawColor(renderer, 255, 0, 0, 255);
+        SDL_RenderDrawRect(renderer, &bounding_box);
+    }
+}
 
 void render_sprite(
     SDL_Renderer* renderer,
@@ -19,12 +42,7 @@ void render_sprite(
 ) {
 
     SDL_Rect source_rect {0, 0, sprite.width_px, sprite.height_px};
-    SDL_Rect dest_rect {
-        (static_cast<int>(transform.position.x) + sprite.horizontal_offset_px) - camera.x, 
-        (static_cast<int>(transform.position.y) + sprite.vertical_offset_px) - camera.y, 
-        sprite.width_px,
-        sprite.height_px
-    };
+    SDL_Rect dest_rect {get_render_target(transform, sprite, camera)};
 
     if (SDL_HasIntersection(&dest_rect, &render_rect)) {
         SDL_RenderCopyEx(
